Consolidate builder namespace
