+++ Confluent.Kafka.DependencyInjection/Clients/ScopedProducer.cs
@@ -1,6 +1,5 @@
 namespace Confluent.Kafka.DependencyInjection.Clients;
 
-using Confluent.Kafka.DependencyInjection.Builders;
 using Confluent.Kafka.DependencyInjection.Handlers;
 
 using Microsoft.Extensions.DependencyInjection;
@@ -31,7 +30,7 @@ class ScopedProducer<TKey, TValue> : IProducer<TKey, TValue>
         IEnumerable<KeyValuePair<string, string>>? config)
     {
        return config != null
-            ? new ProducerAdapter<TKey, TValue>(
+            ? new DIProducerBuilder<TKey, TValue>(
                 new(config),
                 services.GetServices<IErrorHandler>(),
                 services.GetServices<IStatisticsHandler>(),
