+++ Confluent.Kafka.DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,6 +1,5 @@
 namespace Confluent.Kafka.DependencyInjection;
 
-using Confluent.Kafka.DependencyInjection.Builders;
 using Confluent.Kafka.DependencyInjection.Clients;
 using Confluent.Kafka.DependencyInjection.Handlers;
 using Confluent.Kafka.DependencyInjection.Handlers.Default;
@@ -124,8 +123,8 @@ public static class ServiceCollectionExtensions
         services.TryAddSingleton<IOffsetsCommittedHandler, CommitHandler>();
 
         // These must be transient to consume scoped handlers.
-        services.TryAddTransient(typeof(ProducerBuilder<,>), typeof(ProducerAdapter<,>));
-        services.TryAddTransient(typeof(ConsumerBuilder<,>), typeof(ConsumerAdapter<,>));
+        services.TryAddTransient(typeof(ProducerBuilder<,>), typeof(DIProducerBuilder<,>));
+        services.TryAddTransient(typeof(ConsumerBuilder<,>), typeof(DIConsumerBuilder<,>));
 
         return services;
     }
